purchase-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.21.1
pymongo==4.7.0
python-dotenv==1.0.0
pydantic==2.3.0
requests==2.31.0
PyJWT==2.8.0
